#143-3. 리뷰 별점 double로 변경

+++ backend/src/main/java/com/backend/comment/domain/Comment.java
@@ -25,7 +25,7 @@ public class Comment {
 
     private Integer commentCount; // comment의 갯수
     private Integer rateScore; // comment의 별점을 받음
-    private Integer rateScoreAvg; // comment의 별점의 평균
+    private Double rateScoreAvg; // comment의 별점의 평균
 
     private List<File> commentFilesLists;
